api/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.15.2
sqlparse==0.5.5
whitenoise==6.9.0
dj-database-url==2.3.0
psycopg2-binary==2.9.10
Pillow==11.1.0
python-dotenv
requests
